+++ src/vs/base/browser/ui/splitview/paneview.css
