+++ src/vs/base/browser/ui/tree/media/paneviewlet.css
