+++ src/vs/workbench/browser/parts/views/media/paneviewlet.css
