+++ src/vs/workbench/contrib/outline/browser/outline.contribution.ts
@@ -5,7 +5,7 @@
 
 import { localize } from 'vs/nls';
 import { IViewsRegistry, IViewDescriptor, Extensions as ViewExtensions } from 'vs/workbench/common/views';
-import { OutlinePane } from './outlinePanel';
+import { OutlinePane } from './outlinePane';
 import { VIEW_CONTAINER } from 'vs/workbench/contrib/files/common/files';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry';
