+++ src/vs/workbench/contrib/outline/browser/outlinePane.css
